requirements.txt
amulet-core==1.9.27
amulet-leveldb==1.0.2
amulet-nbt==2.1.3
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
lz4==4.3.3
MarkupSafe==3.0.2
marshmallow==3.23.3
mcrcon==0.7.0
mutf8==1.0.6
numpy==1.26.4
packaging==24.2
platformdirs==3.11.0
portalocker==2.10.1
PyMCTranslate==1.2.30
python-dotenv==1.0.1
PyYAML==6.0.2
rcon==2.4.9
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
